pyproject.toml
[project]
name = "aura-discovery"
version = "0.1.0"
description = "Agentic Graph-RAG system for biomedical discovery using Ollama and LangGraph"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "neo4j>=5.18.0,<6.0.0",
    "pydantic>=2.6.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "scikit-learn>=1.4.0",
    "imbalanced-learn>=0.12.0",
    "xgboost>=2.1.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
